fix: additional page reload after renaming
