+++ src/n2-features/f3-profile/InputTypeFile.tsx
@@ -1,16 +1,20 @@
 import React, {ChangeEvent, useState} from 'react';
 import {IconButton} from '@mui/material';
-import defaultAva from '../assets/image/defaultAva.png'
+import defaultAva from '../../assets/image/defaultAva.png'
 import {PhotoCamera} from '@mui/icons-material';
-import {useAppDispatch} from '../n1-main/m2-bll/store';
-import {setAppErrorAC} from '../n1-main/m2-bll/reducers/app-reducer';
+import {useAppDispatch} from '../../n1-main/m2-bll/store';
+import {setAppErrorAC} from '../../n1-main/m2-bll/reducers/app-reducer';
+import {convertFileToBase64} from '../../utils/convertFileToBase64';
 
 type ChangeUserAvatarPropsType = {
     changeUserAvatar: (avatar: string) => void
     userAvatar: string
 }
 
-export const InputTypeFile: React.FC<ChangeUserAvatarPropsType> = React.memo(({changeUserAvatar, userAvatar}) => {
+export const InputTypeFile: React.FC<ChangeUserAvatarPropsType> = React.memo(({
+                                                                                  changeUserAvatar,
+                                                                                  userAvatar,
+                                                                              }) => {
     const dispatch = useAppDispatch()
     const [avatar, setAvatar] = useState<string>(userAvatar ? userAvatar : defaultAva)
     const [isAvatarBroken, setIsAvatarBroken] = useState(false)
@@ -27,15 +31,6 @@ export const InputTypeFile: React.FC<ChangeUserAvatarPropsType> = React.memo(({c
                 dispatch(setAppErrorAC('The file is too large'))
             }
         }
-    }
-
-    const convertFileToBase64 = (file: File, callBack: (value: string) => void) => {
-        const reader = new FileReader();
-        reader.onloadend = () => {
-            const file64 = reader.result as string
-            callBack(file64)
-        }
-        reader.readAsDataURL(file)
     }
 
     const errorHandler = () => {
