+++ src/n2-features/f3-profile/Profile.module.css
@@ -10,7 +10,7 @@
     background: #F9F9FE;
     padding: 20px;
     border-radius: 8px;
-    box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);
+    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
 }
 
 .title {
@@ -23,16 +23,6 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-}
-
-.avatar {
-    /*display: flex;*/
-    /*justify-content: center;*/
-    /*flex-wrap: nowrap;*/
-    border-radius: 20%;
-    border: 1px solid #7b1fa2;
-    width: 120px;
-    height: 120px;
 }
 
 .avatarInput {
